requirements.txt
Django~=2.2.4
django-bootstrap3
misaka
django-braces